requirements.txt
argh==0.26.1
beautifulsoup4==4.4.1
boto==2.38.0
Collectfast==0.2.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.5
django-appconf==1.0.1
django-compressor==1.5
django-constance==1.1.1
django-fresh==1.0.5
django-geoposition==0.2.2
django-grappelli==2.7.1
django-libsass==0.4
django-mailgun==0.8.0
django-picklefield==0.3.2
django-pipeline==1.5.4
django-storages==1.1.5
django-toolbelt==0.0.1
docopt==0.4.0
futures==3.0.3
gunicorn==19.3.0
libsass==0.8.3
mailchimp==2.0.9
mailgun==0.1.1
Markdown==2.6.4
markup==0.2
pathtools==0.1.2
Pillow==3.0.0
psycopg2==2.6.1
pyactiveresource==1.0.1
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.8.1
six==1.10.0
static3==0.6.1
watchdog==0.8.3
whitenoise==2.0.4
flickrapi
python-memcached
stripe